change int to long

+++ src/main/java/au/com/snaker/azcray/data/Maths.java
@@ -2,7 +2,7 @@ package au.com.snaker.azcray.data;
 
 public class Maths {
     // simple factorial number generator
-    public static int generateFactorial(int i) {
+    public static long generateFactorial(long i) {
         if (i == 0) {
             return 1;
         } else {
